+++ tests/testthat/test-cfa.from.keys.R
@@ -19,19 +19,6 @@ test_that(
   }
 )
 test_that(
-  "Test keys not specified",
-  {
-    expect_error(
-      cfa.from.keys(keys = NULL, BFIGritHope, fit_save = FALSE),
-      "one of 'keys_s' or 'keys_e'"
-    )
-    expect_error(
-      cfa.from.keys(data = BFIGritHope, fit_save = FALSE),
-      'argument "keys" is missing'
-    )
-  }
-)
-test_that(
   "No data provided",
   {
     expect_error(
@@ -103,9 +90,9 @@ test_that(
 test_that(
   "Non-list keys list",
   {
-    expect_error(
+    expect_warning(
       cfa.from.keys(keys$grit_c, BFIGritHope, fit_save = FALSE),
-      "'keys_s' is not a list"
+      "'keys' appears to be a vector of items rather than a keys list"
     )
   }
 )