+++ NEWS.md
@@ -1,6 +1,7 @@
 # semFromKeys (development version)
 
 * Fixed a bug which caused `sem.cor` to fail with multi-factor measurement models and items.
+* Entering a vector of items instead of a keys list in `cfa.from.keys` will now work with a warning.
 
 # semFromKeys 0.5.5